Rename modules, packages, and classes from Meshblu to KNoT
