+++ .gitignore
@@ -26,3 +26,6 @@ proguard/
 # Log Files
 *.log
 .idea
+
+#IDE
+.idea/
