--- .idea/modules.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/MeshbluKit-Android.iml" filepath="$PROJECT_DIR$/MeshbluKit-Android.iml" />
-      <module fileurl="file://$PROJECT_DIR$/meshblukit/meshblukit.iml" filepath="$PROJECT_DIR$/meshblukit/meshblukit.iml" />
-    </modules>
-  </component>
-</project>
