+++ knot/src/androidTest/java/com/octoblu/knot/HttpKNoTTest.java
@@ -1,4 +1,4 @@
-package com.octoblu.meshblu;
+package com.octoblu.knot;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
@@ -6,8 +6,8 @@ import android.test.ApplicationTestCase;
 /**
  * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
  */
-public class ApplicationTest extends ApplicationTestCase<Application> {
-    public ApplicationTest() {
+public class HttpKNoTTest extends ApplicationTestCase<Application> {
+    public HttpKNoTTest() {
         super(Application.class);
     }
 }
