+++ knot/src/main/AndroidManifest.xml
