+++ knot/src/main/res/values/strings.xml
