+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':meshblukit'
+include ':knot'
